[Linux] fix method response memory leaks
`fl_method_success_response_new()` does not take ownership of the value
but increments its reference count by one. Using `g_autoptr` is an
"easy" way to ensure that the newly allocated values get unreferenced
at the end of the scope and that way released as appropriate when the
success response is done with them.