Linux: remove misleading C++-style default values
Even if `window_manager_plugin.cc` is compiled as C++, C++-style inline
initialization of struct members doesn't work for `_WindowManagerPlugin`
because the instance is not created by the C++ operator new but alloc'd
and zero-filled by the GObject type system.

> GObject explicitly guarantees that all class and instance members
> (except the fields pointing to the parents) to be set to zero.

https://docs.gtk.org/gobject/concepts.html#object-instantiation

Initial values can be set in `window_manager_plugin_init()` if needed.